Add CRUD to Devices, System & Daemons

+++ mgmt/system.php
@@ -50,6 +50,7 @@
       //   //content += "<span>ID: " + value.id + "</span><br />\n";
       //   console.log( value.mem );
       // });
+//      content += "<span class='sparkline' data-width='100px' size='2.5' color='auto'>3 6 7 4 5 2 3 2 3 4 6</span><br />";
       content += "<span>Host: " + JSON.stringify( systemEntries.data.host ) + "</span><br />";
       content += "<span>Memory: " + JSON.stringify( systemEntries.data.mem ) + "</span><br />";
       content += "<span>Load: " + JSON.stringify( systemEntries.data.load ) + "</span><br />";
@@ -75,14 +76,20 @@
     // shutdown
     
     function systemRead() {
+      let content = "<div> <a href='?action=create'>Create Entry</a><br />\n";
       let entry = apiJson( controllerName, 'read' ).data;
+      if( entry === null || entry.length <= 0 ) {
+        target.html( content + ' No Entries.' );
+        return false;
+      }
       let total = entry.length -1;
-      let content = "<div>";
       $( entry ).each( function( key, value ) {
         content += "<div>";
         content += "<span>ID: " + value.id + "</span><br />\n";
         content += "<span>Name: " + value.name + "</span><br />\n";
-        content += "<span>Data: " + value.data + "</span><br />\n";
+        content += "<span>Value: " + value.value + "</span><br />\n";
+        content += "<span>Updated: " + value.updated + "</span><br />\n";
+        content += "<span>Created: " + value.created + "</span><br />\n";
         content += "<a href='?action=edit&id=" + value.id + "'>Edit</a> <a onclick='return confirm(\"Acknowledge delete Entry " + value.id + " - " + value.name + " - Y/N\")' href='?action=delete&id=" + value.id + "' >Delete</a>";
         content += "<div>";
         if( key < total ) content += "<hr /><br />\n";
@@ -91,37 +98,81 @@
       target.html( content );
     }
     
+    function systemFormFields( item ) {
+      let content = "";
+      if( item === undefined || item.id === undefined || item.id === null ) {
+        item = {
+          name: "",
+          value: "",
+          updated: "",
+          created: ""
+       };
+      } else {
+        content = "<input type='hidden' name='id' value='" + item.id + "' />";
+      }
+      content += "<fieldset>";
+      content += "  <div class='form-group'>";
+      content += "    <label for='name'>Name:</label>";
+      content += "    <input id='name' type='text' name='name' placeholder='" + item.name + "' value='" + item.name + "' class='form-control' />";
+      content += "  </div>";
+      content += "  <div class='form-group'>";
+      content += "    <label for='name'>Value:</label>";
+      content += "    <input id='value' type='text' name='value' placeholder='" + item.value + "' value='" + item.value + "' class='form-control' />";
+      content += "  </div>";
+      content += "  <div class='form-group'>";
+      content += "    <label for='name'>Updated:</label>";
+      content += "    <input id='updated' type='text' name='updated' placeholder='" + item.updated + "' value='" + item.updated + "' class='form-control' />";
+      content += "  </div>";
+      content += "  <div class='form-group'>";
+      content += "    <label for='name'>Created:</label>";
+      content += "    <input id='created' type='text' name='created' placeholder='" + item.created + "' value='" + item.created + "' class='form-control' />";
+      content += "  </div>";
+      content += "  <button type='submit' style='margin:15px 0 0;'>Submit</button> <a href='./" + controllerName + ".php'>Show all</a>";
+      content += "</fieldset>";
+      return content;
+    }
+    
     function systemEdit( id ) {
       //console.log( 'controller: ' + controllerName + ', action: ' + action + ', id: ' + id );
       let entry = apiJson( controllerName, 'read', id ).data;
       let total = entry.length -1;
       let content = "<form id='entryform' action='../' method='POST'>";
+      content += "  <input type='hidden' name='controller' value='" + controllerName + "' />";
+      content += "  <input type='hidden' name='action' value='update' />";
       $( entry ).each( function( key, value ) {
-        content += "<fieldset>";
-//        content += "  <legend>Edit device #" + id + "</legend>";
-        content += "  <input type='hidden' name='controller' value='" + controllerName + "' />";
-        content += "  <input type='hidden' name='action' value='update' />";
-        content += "  <input type='hidden' name='id' value='" + value.id + "' />";
-        content += "  <div class='form-group'>";
-        content += "    <label for='name'>Name:</label>";
-        content += "    <input id='name' type='text' name='name' placeholder='" + value.name + "' value='" + value.name + "' class='form-control' />";
-        content += "  </div>";
-        content += "  <div class='form-group'>";
-        content += "    <label for='name'>Data:</label>";
-        content += "    <input id='data' type='text' name='data' placeholder='" + value.data + "' value='" + value.data + "' class='form-control' />";
-        content += "  </div>";
-        content += "  <button type='submit' style='margin:15px 0 0;'>Submit</button>"
-        content += "</fieldset>";
+        content += systemFormFields( value );
       });
-      content += "<form><br />";
-      content += "<a href='./" + controllerName + ".php'>Overview</a>";
+      content += "</form><br />";
       target.html( content );
     }
     
     function systemDelete( id ) {
       let result = apiJson( controllerName, 'delete', id );
+      target.html( result );
       return result;
     }
+    
+    function systemCreate() {
+      //console.log( 'controller: ' + controllerName + ', action: ' + action + ', id: ' + id );
+      //let devices = currentDaemons;
+      let systems = {
+        id: null,
+        name: "",
+        value: "",
+        updated: "",
+        created: ""
+      };
+      let content = "<form id='devicesform' action='../' method='POST'>";
+      content += "  <input type='hidden' name='controller' value='" + controllerName + "' />";
+      content += "  <input type='hidden' name='action' value='create' />";
+      $( systems ).each( function( key, value ) {
+//        currentType = value.type;
+        content += systemFormFields( value );
+      });
+      content += "</form><br />";
+      target.html( content );
+    }
+    
     
     // Main caller
     $(document).ready( function() {
@@ -147,6 +198,8 @@
           systemDashboard();
           systemRead();
         }
+      } else if( action === 'create' ) {
+        systemCreate( id );
       }
     });
